+++ context_use/providers/airbnb/search_history/__init__.py
@@ -1,0 +1,3 @@
+from context_use.providers.airbnb.search_history.pipe import AirbnbSearchHistoryPipe
+
+__all__ = ["AirbnbSearchHistoryPipe"]